add logging and regex for customer

+++ src/main/java/com/example/posbackendspring/controller/CustomerController.java
@@ -1,31 +1,36 @@
 package com.example.posbackendspring.controller;
 
 import com.example.posbackendspring.dto.impl.CustomerDTO;
+import com.example.posbackendspring.exception.CustomerNotFoundException;
 import com.example.posbackendspring.exception.DataPersistException;
-import com.example.posbackendspring.exception.NoteNotFoundException;
 import com.example.posbackendspring.service.CustomerService;
+import com.example.posbackendspring.util.RegEx;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.regex.Pattern;
-
 @RestController
 @RequestMapping("api/v1/customers")
 public class CustomerController {
     @Autowired
     private CustomerService customerService;
+    private static Logger logger = LoggerFactory.getLogger(CustomerController.class);
 
     @PostMapping (consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Void> saveCustomer(@RequestBody CustomerDTO customerDTO) {
         try {
             customerService.saveCustomer(customerDTO);
+            logger.info("Customer Save Successfully");
             return new ResponseEntity<>(HttpStatus.CREATED);
         } catch (DataPersistException e) {
+            logger.warn("Returning Http 400 Bad Request",e.getMessage());
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         } catch (Exception e) {
+            logger.error("Customer save unsuccessful",e.getMessage());
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
@@ -33,17 +38,39 @@ public class CustomerController {
 
     @DeleteMapping(value = "/{customerId}")
     public ResponseEntity<Void> deleteCustomer(@PathVariable("customerId") String customerId) {
+        try {
+            if (!RegEx.customerIdMatcher(customerId)) {
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
+            logger.info("Customer delete successful");
             customerService.deleteCustomer(customerId);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }catch (CustomerNotFoundException e) {
+            logger.warn("Returning Http 400 Bad Request",e.getMessage());
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }catch (Exception e){
+            logger.error("Customer Delete unsuccessful",e.getMessage());
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
     @PutMapping(value = "/{customerId}")
     public ResponseEntity<Void> updateCustomer(@PathVariable("customerId") String customerId,
                                            @RequestBody CustomerDTO updatedCustomerDto) {
         //validation
+        try {
+            if (!RegEx.customerIdMatcher(customerId) || updatedCustomerDto == null) {
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
             customerService.updateCustomer(customerId, updatedCustomerDto);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } catch (CustomerNotFoundException e) {
+            logger.warn("Returning Http 400 Bad Request", e.getMessage());
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        } catch (Exception e) {
+            logger.error("Customer update unsuccessful", e.getMessage());
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
-
 
     @GetMapping(value = "/{customerId}",produces = MediaType.APPLICATION_JSON_VALUE)
     public CustomerDTO getCustomer(@PathVariable("customerId") String customerId){
